refactor: ContactManager 제거 및 NotificationCenter 제거
- ContactManager 제거 후 ContactMangerVC의 Controller 역할 재배정
- NotificationCenter 제거 후 delegate 패턴으로 변경
