+++ ContactManager/ContactManager/Controller/ContactChangedDelegate.swift
@@ -1,0 +1,10 @@
+//
+//  ContactChangedDelegate.swift
+//  ContactManager
+//
+//  Created by 김예준 on 10/17/23.
+//
+
+protocol ContactChangedDelegate {
+  func reload()
+}
